Weekly Graded Assignment for Chapter 13
The starter code below contains a Point class. Add a method slopeFromOrigin, which returns the slope of the line joining the origin to the point. For example:
>>> Point(4, 10).slopeFromOrigin()
2.5
>>> Point(12, -3).slopeFromOrigin()
-0.25
>>> Point(-6, 0).slopeFromOrigin()
0
The equation for calculating slope between any two points is slope = (Y2 - Y1) / (X2 - X1). Remember that dividing by 0 is not allowed, so there are some inputs that will cause your method to fail. Return None when that happens.